Move functions to platform specific files
